add leveldb support
